requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
requests
python-dotenv
httpx
pytest
